requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
websockets==13.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# Caching
redis==5.2.0
hiredis==3.0.0

# AI/ML
google-generativeai==0.8.3
langdetect==1.0.9

# Audio Processing
pydub==0.25.1
gtts==2.5.3

# Image Processing
Pillow==11.0.0

# Communication
twilio==9.3.7

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==44.0.0
python-dotenv==1.0.1

# Utilities
pydantic==2.10.2
pydantic-settings==2.6.1
requests==2.32.3
aiohttp==3.11.8
httpx==0.27.2

# Monitoring
prometheus-client==0.21.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# Development
black==24.10.0
flake8==7.1.1
mypy==1.13.0
psutil==6.1.0
